+++ src/components/Cards/index.ts
@@ -1,0 +1,7 @@
+import { LoginCard } from "./LoginCard";
+import { SignupCard } from "./SignupCard";
+
+export {
+    LoginCard,
+    SignupCard
+}
