+++ src/components/Cards/loginsignup.scss
@@ -1,0 +1,5 @@
+.login-signup-body {
+    display: flex;
+    justify-content: center;
+    padding: 0 0 size(5) 0;
+}
