+++ src/components/Main/LandingLayout.tsx
@@ -1,15 +1,6 @@
 import React from "react";
-import { Main, MainBody } from ".";
-import AlertMessage from "../AlertMessage";
-import { Block, BlockBody, BlockFooter, BlockHeader } from "../Block";
-import Breadcrumb from "../Breadcrumb";
-import Button from "../Button";
 import { Footer, FooterList, FooterCopyright } from "../Footer";
 import Gradient from "../Gradient";
-import InputBox from "../InputBox";
-import { Navigation, NavigationGroup, NavigationItem, NavigationSection } from "../Navigation";
-import Profilebar from "../Profilebar";
-import Sidebar from "../Sidebar";
 import "./main.scss";
 
 const Copyright = {
