+++ src/components/Main/Main.stories.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import { Meta } from "@storybook/react/types-6-0";
 import { Story } from "@storybook/react";
 import { MainPage } from './MainPage';
-import { LoginPage } from './LoginPage';
 import { LandingLayout } from './LandingLayout';
-import { LoginCard } from './LoginCard';
-import { SignupCard } from './SignupCard';
+import { LoginCard } from '../Cards';
+import { SignupCard } from '../Cards';
 export default {
     title: "Components/Main",
 } as Meta;
