+++ src/components/Main/main.scss
@@ -28,12 +28,6 @@
             text-align: center;
         }
     }
-
-    .login-body {
-        display: flex;
-        justify-content: center;
-        padding: 0 0 size(5) 0;
-    }
 }
 
 .main {
